pyproject.toml
[project]
name = "c4"
version = "0.10.0"
description = "C4 Python Sidecar - LSP, Knowledge, GPU services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "sqlite-vec>=0.1.0",
    "jedi>=0.19.0",
    "multilspy>=0.0.15",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-javascript>=0.25.0",
    "tree-sitter-typescript>=0.23.2",
]

[project.optional-dependencies]
embeddings = [
    "openai>=1.0.0",
    "sentence-transformers>=2.0.0",
]
gpu = [
    "pynvml>=11.0.0",
    "psutil>=5.9.0",
]
review = [
    "anthropic>=0.40.0",
    "openai>=2.17.0",
    "numpy>=2.2.6",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
]
document = [
    "pymupdf>=1.24.0",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "openpyxl>=3.1.5",
    "pdfplumber>=0.11.9",
    "xlrd>=2.0.2",
    "python-hwpx>=1.9",
]

[project.scripts]
c4-bridge = "c4.bridge.sidecar:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["c4"]

[tool.ruff]
line-length = 140
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
]
